Fix crash when -O0 because of no stack.
The reason for the crash is because when seL4 invokes an application
there is no stack. So I've added min_start.S with a _min_start entry
point.

I then update Makefile and main which can not return because min_start
is handling "exit", which for now is just hanging in an endless loop.